+++ packages/experience-builder-sdk/src/core/sdkFeatures.ts
@@ -1,0 +1,2 @@
+// Object to store the SDK features that are enabled/disabled
+export const sdkFeatures: Record<string, unknown> = {};